index.py
import asyncio
import yaml

# Load YAML configuration
def load_config(config_file):
    with open(config_file, 'r') as yaml_file:
        config = yaml.safe_load(yaml_file)
    return config

# Asyncflows backend functions (placeholder implementations)
async def async_parse_resume(resume_content):
    # Placeholder function for resume parsing
    print(f"Parsing resume: {resume_content}")
    parsed_resume = f"Parsed resume: {resume_content[:50]}..."
    return parsed_resume

async def async_generate_cover_letter(parsed_resume, job_description, config):
    # Placeholder function for cover letter generation
    print(f"Generating cover letter for job '{job_description}' based on resume '{parsed_resume}'")
    cover_letter = (
        f"{config['cover_letter_generator']['greeting']}\n\n"
        f"Dear Hiring Manager,\n"
        f"Based on your job description for {job_description}, "
        f"I believe my skills in {parsed_resume} make me a strong candidate...\n\n"
        f"{config['cover_letter_generator']['closing']}"
    )
    return cover_letter

# Example usage for testing purposes
if __name__ == "__main__":
    config = load_config('index.yaml')

    # Asyncflows tasks based on YAML flow definition
    async def run_flow(resume_content, job_description):
        parsed_resume = await async_parse_resume(resume_content)
        cover_letter = await async_generate_cover_letter(parsed_resume, job_description, config)
        return cover_letter

    # Example usage (replace with actual usage in your application)
    async def example_async_operation():
        resume_content = "John Doe\n123 Main St\nPython Developer"
        job_description = "Python Developer"
        result = await run_flow(resume_content, job_description)
        print(result)

    asyncio.run(example_async_operation())
