requirements.txt
asyncflows
gradio
pyyaml